Migrating downsample with security yaml IT tests
