+++ x-pack/plugin/downsample/qa/with-security/build.gradle
@@ -5,13 +5,11 @@
  * 2.0.
  */
 
-import org.elasticsearch.gradle.Version
-import org.elasticsearch.gradle.internal.info.BuildParams
-
-apply plugin: 'elasticsearch.legacy-yaml-rest-test'
-apply plugin: 'elasticsearch.legacy-yaml-rest-compat-test'
+apply plugin: 'elasticsearch.internal-yaml-rest-test'
+apply plugin: 'elasticsearch.yaml-rest-compat-test'
 
 dependencies {
+  testImplementation project(path: ':test:test-clusters')
   yamlRestTestImplementation project(path: xpackModule('rollup'))
 }
 
@@ -21,11 +19,11 @@ restResources {
   }
 }
 
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  setting 'xpack.license.self_generated.type', 'trial'
-  setting 'xpack.security.enabled', 'true'
-  user username: 'elastic_admin', password: 'admin-password'
+tasks.named('yamlRestTest') {
+  usesDefaultDistribution("to be triaged")
+}
+tasks.named('yamlRestCompatTest') {
+  usesDefaultDistribution("to be triaged")
 }
 
 if (buildParams.inFipsJvm){
